Dele Beregne.tsx i BP og OMS
